requirements.txt
websocket-client
Pillow
streamlit
st-clickable-images
requests-toolbelt
